Remove non intrusive incall ui (3/3)
There are several reasons for it

- to call every time a complete new activity is the wrong way. It needs
to be handled like google is doing it which keeps the activity permanent
alive to get it as fast as possible to the front. Especially noticeable on
slower phones

- of course current way freezes the forground app....so question what is
non intrusive on this feature

- Handling what happens when screen if off and the user is not in an
activity needs to be reworked

Whatever currently we have not much time to work that all out. So for
now we remove till this feature is completly redone. All points are solvable
but need work especially point 2 which is the most work. Other points are easy
to solve. Aside that our old feature "call in background"
returns back.

Change-Id: I629b749927d154cd36438b71c4a3552a7d6e3bf6

Conflicts:
	res/layout/card_call_incoming.xml
	src/com/android/incallui/InCallCardActivity.java