InCallUI: add decline and answer  incoming call action on notifications
Add decline and answer call action on notification if an call is incoming. Especially
usefull for call in background feature to have the ability to directly
decline or answer a incoming call. But as well usefull to decline/answer a second call
during one is incomming.

PS:
- found a better way for the answer button

PS:
- fix keep flip action alive
- fix collapse notification shade when answering call
- answer call before starting the UI to do not have
  shown the incall screen for a brief second

PS:
- fix wrong user on sending close system dialogs

Change-Id: I92d8f5a2de74f9877a61c4da4c3eaf2e25230895

Conflicts:
	src/com/android/incallui/InCallPresenter.java